requirements.txt
numpy==1.24.4
scipy==1.10.1
pyyaml==6.0.1
ray==1.13.0
toolz==0.12.1
-i https://download.pytorch.org/whl/cu117
torch_scatter==2.1.2
torch_geometric==2.5.3
seaborn==0.13.2